fix: PR #520
