+++ core/api-server/lib/service/webhooks.js
@@ -25,7 +25,7 @@ class WebhooksService {
     async getWebhooks(options) {
         validator.validateJobID(options);
         const webhooks = await stateManager.getWebhooks({ jobId: options.jobId });
-        if (!webhooks) {
+        if (webhooks.length === 0) {
             throw new ResourceNotFoundError('webhook', options.jobId);
         }
         const { progress, result } = webhooks[0];
