+++ core/api-server/tests/init.js
@@ -1,0 +1,21 @@
+const bootstrap = require('../bootstrap');
+const stateManager = require('../lib/state/state-manager');
+const { algorithms, pipelines, webhookStub } = require('./mocks');
+
+before(async () => {
+    const config = await bootstrap.init();
+    const baseUrl = `${config.swagger.protocol}://${config.swagger.host}:${config.swagger.port}`;
+    await Promise.all(pipelines.map(p => stateManager.setPipeline(p)));
+    await Promise.all(algorithms.map(p => stateManager.setAlgorithm(p)));
+    webhookStub.start();
+
+    const restUrl = `${baseUrl}/${config.rest.prefix}/v1`;
+    const internalUrl = `${baseUrl}/internal/v1`;
+
+    global.testParams = {
+        config,
+        baseUrl,
+        restUrl,
+        internalUrl
+    }
+});
