+++ core/api-server/tests/mocha.opts
@@ -1,0 +1,1 @@
+--file ./init
